serialize LedgerIndex using longValue instead of intValue

+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/model/jackson/modules/LedgerIndexSerializer.java
@@ -41,7 +41,7 @@ public class LedgerIndexSerializer extends StdScalarSerializer<LedgerIndex> {
 
   @Override
   public void serialize(LedgerIndex ledgerIndex, JsonGenerator gen, SerializerProvider provider) throws IOException {
-    gen.writeNumber(ledgerIndex.unsignedIntegerValue().intValue());
+    gen.writeNumber(ledgerIndex.unsignedIntegerValue().longValue());
   }
 
 }
